add Upstart 1.8 file
:Release Notes:
- Move the Upstart 1.8 job file from starfish-initscripts.
- see GF-8224 for the install for the new job file

:Detailed Notes:
The intent of this fix is moving the upstart script - in the
initial implementation of Upstart 1.8, many 'component' upstart jobs
were modified for 1.8 and packaged with starfish-initscripts to ease
the transition from Upstart 0.3.11 to 1.8.  The plan was and is to
move the 'imported' scripts back to their home package as soon as is
reasonable.
This particular fix is tied to the update to
meta-webos/recipes-webos/app-services/app-services.bb that will perform
the install of the new file.

This fix MUST be merged before or at the same time as that one or the
build will break.
The meta-webos change is [GF-8224] and is covered by
https://g2g.palm.com/#/c/2792/

:Testing Performed:
Since the 1.8 file itself is functionally unchanged from that present
in starfish-initscripts, the testing done running Upstart 1.8 + s-i
over the last couple of months still applies.

The testing done specifically for this fix is to locally build the
component and inspect the output directory:
starfish/BUILD-goldfinger/work/all-starfish-linux/app-services/
   3.0.0-1-r6starfish1/image/etc/init and ..../etc/event.d
to confirm that the properly configured upstart scripts end up in their
respective locations.

:QA Notes:

:Issues Addressed:
[GF-8223] add Upstart 1.8 script to app-services
[GF-8224] update meta-webos/recipes-webos/app-services/app-services.bb to install upstart 1.8 script

Open-webOS-DCO-1.0-Signed-off-by: Ed Chejlava <ed.chejlava@lge.com>

Change-Id: I33c09a88c86f0c9c2913852d7d6ae5fe401eb61e
Reviewed-on: https://g2g.palm.com/2788
Reviewed-by: Build Verification
Reviewed-by: Ed Chejlava <ed.chejlava@lge.com>
Tested-by: Ed Chejlava <ed.chejlava@lge.com>
Reviewed-by: Suresh Arumugam <suresh.arumugam@lge.com>
Tested-by: Suresh Arumugam <suresh.arumugam@lge.com>